requirements.txt
annotated-types==0.7.0
eralchemy==1.5.0
greenlet==3.2.2
langdetect==1.0.9
mysql-connector-python==9.3.0
numpy==2.2.6
pandas==2.2.3
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
dotenv